+++ internal/runners/flags.go
@@ -1,4 +1,4 @@
-package register
+package runners
 
 import (
 	"fmt"
